quant: preserve caller dtype in fake-quant round-trip
Previously the int{2,4,8} pack/unpack and FP8 quant/dequant helpers cast their
inputs to torch.float16 unconditionally. With a bf16 model (Qwen3, Llama-3,
Mistral-7B, etc.), every round-trip through the fake-quant lost precision
twice — bf16 -> fp16 on entry, fp16 -> bf16 on exit — on top of the actual
quantization error we want to measure.

This makes the helpers dtype-agnostic by routing through the input/scale
dtype instead of hard-coded fp16:
- quant/new_pack.py::unpack_and_dequant_{k,v}cache: data.to(scale.dtype)
- quant/new_pack.py::dequantize_cache_pertoken: out.to(scale_exp.dtype)
- quant/fp8_quant.py: scale stored at data.dtype; dequant uses scale.dtype
- vllm_custom/fake_quant_utils.py: drop the .to(torch.float16) coercions
  inside fake_quantize_{fp8, k_perchannel, v_pertoken}; the round-trip stays
  in the caller's dtype and only re-casts to orig_dtype defensively at the
  outer boundary.

+++ quant/fp8_quant.py
@@ -45,7 +45,7 @@ def quantize_fp8(
     data_fp8 = data_scaled.to(_FP8_DTYPE).view(B, nh, T, D)
     data_uint8 = data_fp8.view(torch.uint8)
 
-    return data_uint8, scale.to(torch.float16)
+    return data_uint8, scale.to(data.dtype)   # match caller dtype (bf16/fp16)
 
 
 def dequantize_fp8(
@@ -70,6 +70,6 @@ def dequantize_fp8(
     # Expand scale from (B, nh, T, num_groups) -> (B, nh, T, D)
     scale_expanded = scale.unsqueeze(-1).expand(B, nh, T, num_groups, group_size).reshape(B, nh, T, D)
 
-    out = data_fp8.to(torch.float16) * scale_expanded
+    out = data_fp8.to(scale.dtype) * scale_expanded
     # FP8 e4m3fn can carry NaN; guard downstream attention math.
     return torch.nan_to_num(out, nan=0.0, posinf=0.0, neginf=0.0)
